+++ ray tracing/ray tracing/material.h
@@ -34,7 +34,7 @@ public:
 
 class metal : public material {
 public:
-    metal(const color& a, double f) : albedo(a), fuzz(f < 1 ? f : 1) {}
+    metal(const color& a, float f) : albedo(a), fuzz(f < 1 ? f : 1) {}
 
     virtual bool scatter(
         const ray& r_in, const hit_record& rec, color& attenuation, ray& scattered
@@ -47,22 +47,22 @@ public:
 
 public:
     color albedo;
-    double fuzz;
+    float fuzz;
 };
 
 class dielectric : public material {
 public:
-    dielectric(double index_of_refraction) : ir(index_of_refraction) {}
+    dielectric(float index_of_refraction) : ir(index_of_refraction) {}
 
     virtual bool scatter(
         const ray& r_in, const hit_record& rec, color& attenuation, ray& scattered
     ) const override {
         attenuation = color(1.0, 1.0, 1.0);
-        double refraction_ratio = rec.front_face ? (1.0 / ir) : ir;
+        float refraction_ratio = rec.front_face ? (1.0 / ir) : ir;
 
         vec3 unit_direction = vec3::unit_vector(r_in.direction());
-        double cos_theta = fmin(vec3::dot(-unit_direction, rec.normal), 1.0);
-        double sin_theta = sqrt(1.0 - cos_theta * cos_theta);
+        float cos_theta = fmin(vec3::dot(-unit_direction, rec.normal), 1.0);
+        float sin_theta = sqrt(1.0 - cos_theta * cos_theta);
 
         bool cannot_refract = refraction_ratio * sin_theta > 1.0;
         vec3 direction;
@@ -76,10 +76,10 @@ public:
     }
 
 public:
-    double ir; // Index of Refraction
+    float ir; // Index of Refraction
 
 private:
-    static double reflectance(double cosine, double ref_idx) {
+    static float reflectance(float cosine, float ref_idx) {
         // Use Schlick's approximation for reflectance.
         auto r0 = (1 - ref_idx) / (1 + ref_idx);
         r0 = r0 * r0;
