+++ ray tracing/ray tracing/ray.h
@@ -16,5 +16,5 @@ public:
     point3 origin() const;
     vec3 direction() const;
 
-    point3 at(double t) const;
+    point3 at(float t) const;
 };
